### 2021-04-13 #3161 @will-rowe
Hi @smjenness - would you be willing to review this submission to JOSS?
