### 2021-04-14 #3161 @smjenness
The "manuscript" (really just an extended abstract) appears fine. I am not able to review the software itself, as I don't work in Python.
