### 2021-04-19 #3161 @GISKid
This loooks really important. I do not have capacity at this time.

On Thu., Apr. 15, 2021, 8:52 a.m. Will Rowe, ***@***.***>
wrote:

> Hi @GISKid <https://github.com/GISKid> - would you be able to review this
> JOSS submission?
>
