### 2021-05-04 #3161 @joeweaver
Hi Will,

I'd be a first time reviewer for JOSS, so thanks for having a review
checklist publicly available so I know what I'd be responsible for
evaluating. I'd be happy to review and would almost certainly have time
within the next two weeks to do so (likely much sooner, but cannot
guarantee).

I can say, based on a cursory look at the git repo, I'd have some issues
with the minimal documentation.

--Joe
