### 2021-05-04 #3161 @mjacob1002
Hi @joeweaver,

Thanks for the feedback! Just to clarify for myself: should I move the docstrings in the files python files themselves to the "docs" folder or add more to the docstrings themselves?

Best,
Mathew
