### 2021-05-04 #3247 @whedon
[REVIEW]: Eir: A Python Package for Epidemic Simulation
**Submitting author:** @mjacob1002 (<a href="http://orcid.org/0000-0001-5513-432X">Mathew Jacob</a>)
**Repository:** <a href="https://github.com/mjacob1002/Eir" target ="_blank">https://github.com/mjacob1002/Eir</a>
**Version:** v1.0.2
