### 2021-05-04 #3161 @joeweaver
The docstrings look ok. Probably no need to move all the variable
explanations over (no need to put the whole API into the paper, which I
gather is counter to JOSS), but maybe do so for the example in the README
file.  The goal (at least from the perspective of a reviewer coming from
outside epidemiology) is to provide just a touch more context.

On Tue, May 4, 2021 at 3:48 PM Mathew Jacob ***@***.***>
wrote:

> Hi @joeweaver <https://github.com/joeweaver>,
>
> Thanks for the feedback! Just to clarify for myself: should I move the
> docstrings in the files python files themselves to the "docs" folder or add
> more to the docstrings themselves?
>
> Best,
> Mathew
>
