### 2021-05-11 ce1d553
Added the variable explanations of the Periodic Mobility Model directly to the README.md file
