### 2021-05-11 607158f
Finished adding details for documentation of the example in the README
