+++ README.md
@@ -39,9 +39,9 @@ If one were to model the ICU hospitalizations using the Hub Model, the code coul
 ```python
 from Eir import PeriodicICUV
 
-  test = PeriodicICUV(S0=999, E0=0, I0=1, R0=0, V0=0, rho=.3, ioda=.3, gamma=.25, mu=0.007, omega=.14, phi = .42, chi=.15, kappa=.05, eta=.02, spread_r=2, sigma_r=.25, move_R=4, sigma_R=.75, side=33, days=31)       
-  test.run()
-  test.plot()
+test = PeriodicICUV(S0=999, E0=0, I0=1, R0=0, V0=0, rho=.3, ioda=.3, gamma=.25, mu=0.007, omega=.14, phi = .42, chi=.15, kappa=.05, eta=.02, spread_r=2, sigma_r=.25, move_R=4, sigma_R=.75, side=33, days=31)       
+test.run()
+test.plot()
 ```
 In the above code segment:
   S0 : int
