### 2021-05-11 #3247 @mjacob1002
Hi @joeweaver and @DARSakthi,

Thank you for reviewing my JOSS submission!

@joeweaver I added documentation for the README file example - let me know if you think there is anything else I should do in terms of documentation.

Thanks again!
- Mathew Jacob
