### 2021-05-14 #3247 @joeweaver
Hi Matthew,

Looking better.  I intend to review more deeply early next week.

--Joe

On Tue, May 11, 2021 at 11:04 PM Mathew Jacob ***@***.***>
wrote:

> Hi @joeweaver <https://github.com/joeweaver> and @DARSakthi
> <https://github.com/DARSakthi>,
>
> Thank you for reviewing my JOSS submission!
>
> @joeweaver <https://github.com/joeweaver> I added documentation for the
> README file example - let me know if you think there is anything else I
> should do in terms of documentation.
>
> Thanks again!
>
>    - Mathew Jacob
>
