### 2021-05-18 #3247 @joeweaver
Hi @will-rowe,

I'm starting to review this in earnest. It's my first time interacting with JOSS or @whedon and it seems there's a bit of a workflow for me to learn.

Do I just leave my my checklist and any additional comments in this issue, or is there some other process?
