### 2021-05-18 #3247 @will-rowe
Hi @joeweaver

Thanks again for agreeing to review - looking forward to hearing what you think of this submission.

We recommend creating one or more issues in the [repository being reviewed](https://github.com/mjacob1002/Eir), either one for all your comments/requests or you can separate out over multiple issues if that is clearer. You can then reference those issues back here in the JOSS review issue so myself and others can track progress. Once you are happy that each item in your checklist has been satisfied, please tick them off here and ping me to say you are done.

Any other questions just let me know. There are also some guidelines [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html)
