### 2021-05-18 mjacob1002/Eir#9 @joeweaver
Paper Review: Unit test documentation
Minor edit, but it would be nice to have a brief (1-2 sentences) on how to kick off the unit tests.

It looks like you're comparing results against pre-generated CSVs.  Are the data in those CSVs a 'gold standard' for these models, or are they to prevent regressions within your own implementations?  Either is fine, I'd just prefer to have that documented somewhere.
