### 2021-05-18 #3247 @joeweaver
It's working now. The re-invite didn't originally fix it, but turning off my adblocker & Firefox enhanced tracking protection allowed it through (or it was a coincidence).
