### 2021-05-18 mjacob1002/Eir#10 @joeweaver
Paper Review: Related Packages/State of the Field
There's a list of other packages (which is great!), but there's no explicity comparison between EIR and these. No need to do big analysis or feature-bullet chart.  Just something on the order of a short paragraph that helps me see how EIR fits into the ecosystem.
