### 2021-05-18 mjacob1002/Eir#9 @mjacob1002
Hi @joeweaver,

Because of the stochasticity of the package, in order to have consistency, I seeded the numpy RNG and then documented the outputs of that particular seed and wrote it to a CSV file. This would allow for there to be a correct output to check against. I'll document it in the docs folder.
