### 2021-05-19 #3247 @DARSakthi
Hi,

The author presents a centralised package to automate the selection and simulation of various dynamical models of disease transmission. It is novel, if incremental, in that it comes packaged; more importantly, from what I can discern, it is a useful contribution to the methods in this field. I think my review is mostly complete, so I'll proceed with it here

Comments on the code:

Code is clean. Good docs and performance. The simulations worked well and ran to completions with no errors. I did all tests and a few 'spot-checks.'

1.1 - The axis formatting on some plots is not great, and I thought it might be nice for the timeseries visualisation to have more 'bells and whistles,' rather than the standard pyplot type of thing. But, the former could be my display, and the latter is not crucial, so it shouldn't bar the paper from publication. Something to consider for the (very) near future, perhaps.

1.2 - SEIR does not specify that run() takes parameters. This should be stated in docs

1.3 - SIRS documentation is incomplete, as it does not include beta. Also does not specify that run takes parameters

1.4 - I guess SIS is a class and isn't meant to run. If my assumption is an error, then SIS doesn't run. If I'm right, please specify this.

Comments on the paper:

2.1 - I find the summary uncompelling. There is a good review of the state of the art, but from the language provided alone, it is unclear what the software package itself does.

2.2 - In point of fact, it might be best to switch the Summary and Statement of Need section, because I have the opposite criticism of the Statement of Need. Now I know what the software does, but it is unclear what the use of it is. The language in a few of the repository's README files are good for this; you may consider lifting it into the paper to convey the point of the package, or what it aims to do.

I recommend cleaning up both sections to adhere to the guidelines provided by JOSS. Note that these are defined as
* A summary _describing the high-level functionality_ and _purpose_ of the software for a diverse, non-specialist audience.
* A Statement of Need section that _clearly_ illustrates the _research purpose_ of the software.

Switching would be recommended, but might not be sufficient to satisfy these guidelines. It is advisable that some refactoring is performed to communicate these points more clearly. The current Summary section (recommended to be changed to Statement of Need) is what I refer to here -- it needs more elaboration on what _motivates_ this package. Changing Statement of Need to the Summary is sufficient for a satisfactory new Summary section. You may also wish to mix portions of each section -- I'll leave it up to you -- so long as the above two points are satisfied. Currently I don't think they are

2.3 - I would like to see more details about what sets this package apart from the two indicated in section "Related Packages." Please comment (briefly is fine) on what is done here that is not done there.

2.4 - line 42, "Eir is currently being used for a publication involving simulating hospitalizations and studying vaccinations strategies." If such an article is available, either as a preprint or post-review publication, I suggest citing it -- it would satisfy an obligation to report relevant literature and document use-cases (not to mention looking nice for the author)

2.5 - lines 43-45, " It is intended to be used by epidemiologists and researchers who are interested in understanding the dynamics of epidemics _when operating under certain assumptions._" Like what? Can you indicate examples, e.g., the variables in some models?

2.6 - There is a typo on line 62, 'guiding.'

2.7 - It seems strange to acknowledge a lack of funding. Can the editor @will-rowe comment on this from the standpoint of JOSS guidelines? Is it better to keep it in as a simple statement, or to get rid of it? It might be that it doesn't really matter.

Overall, I would recommend acceptance, given the revisions outlined above.

Cheers,

Dalton A R Sakthivadivel
