### 2021-05-20 c477d07
Added shell files that automate tests for each type of file, as well as a master test file that runs everything.
