+++ tests/Periodic/test_Periodic.sh
@@ -1,0 +1,3 @@
+echo "Running test for Periodic Models:";
+for FILE in `ls *.py`; do echo -e $FILE; python3 $FILE; done
+
