+++ tests/randMove/test_randMove.sh
@@ -1,0 +1,3 @@
+echo "Running test for Random Movement Models:";
+for FILE in *.py; do echo -e $FILE; python3 $FILE; done
+
