+++ tests/Deterministic/test_Deterministic.sh
@@ -1,0 +1,3 @@
+echo "Running test for Deterministic Models:";
+for FILE in `ls *.py`; do echo -e $FILE; python3 $FILE; done
+
