+++ tests/Hub/test_Hub.sh
@@ -1,0 +1,3 @@
+echo "Running test for Static Hub Models:";
+for FILE in *.py; do echo -e $FILE; python3 $FILE; done
+
