+++ docs/unitTest/how_to_run.md
@@ -1,0 +1,14 @@
+In order to run each test, after forking the repository, go to each of the python file for each of the tests and run the following in the shell:
+
+```shell
+
+python3 <NAME_OF_FILE>
+```
+
+For example, if I wanted to run the test for the HubSEIRSV model, the following command would be run:
+
+```shell
+
+python3 test_HubSEIR.py
+
+```
