+++ docs/unitTest/test_details.md
@@ -1,0 +1,4 @@
+# Structure
+
+Because Eir is an epidemiological simulation/modeling package, the outputs are stochastic. Therefore, to control for this, the random number generator in each of the tests are seeded. Therefore, the RNG will produce the same numbers, which should therefore lead to reproducible results. The CSVs contain the outputs that each test file should output with the seed they have. 
+
