+++ docs/unitTest/how_to_run.md
@@ -12,3 +12,14 @@ For example, if I wanted to run the test for the HubSEIRSV model, the following 
 python3 test_HubSEIR.py
 
 ```
+Additionally, the shell files in each subdirectory(Deterministic, Hub, Periodic). For example, to run all of the periodic model tests from the tests/Periodic/ directory:
+
+```shell
+
+./test_Periodic.sh
+```
+
+In order to test every test for every model type, run the following from the tests/ directory:
+```shell
+./all_tests.sh
+```
