### 2021-05-20 658a19d
Fixed type in paper.md
