### 2021-05-20 7d9e8f7
Switched Statement of Need and Summary. Also briefly added to the Statement of Need section of paper.
