+++ docs/Deterministic/SIRD.md
@@ -10,7 +10,7 @@ sim = SIRD(S0=9999999, I0=1, R0=0, beta=1.5, gamma=.15, omega=.01)
 df = sim.run(31, .1, plot=False)
 ```
 
-run() takes the number of simulated as first parameter, and the differential, or step, used for the approximation of the ODE.
+run() takes the number of simulated as first parameter, and the differential, or step, used for the Euler approximation of the ODE.
 
 If a plot is wanted, simply run the following:
 
