### 2021-05-24 #3247 @will-rowe
Thanks for a thorough review @DARSakthi. It looks like @joeweaver has completed his as well (issues [here](https://github.com/mjacob1002/Eir/issues/9) and [here](https://github.com/mjacob1002/Eir/issues/10))? Please confirm @joeweaver.

@mjacob1002 - there are a few boxes yet to be checked in both reviews, please could you work toward achieving these. In addition please also respond to the specific comments raised by @DARSakthi and @joeweaver.

Thanks all
