### 2021-05-24 75326e6
Fixed describing how Eir fits into the ecosystem in the paper.
