### 2021-05-24 553e97d
Delete paper.md
