### 2021-05-24 mjacob1002/Eir#10 @mjacob1002
Hi @joeweaver! Sorry for the delay - it's exam season at the moment! I think I addressed the issue by stating how Eir was different from each of the packages when I mentioned them. Let me know if that fixes the issue.
