+++ docs/Deterministic/SEIR.md
@@ -10,7 +10,7 @@ sim = SEIR(S0=9999999, I0=1, R0=0, beta=1.5, rho=.25, gamma=.15)
 df = sim.run(31, .1, plot=False)
 ```
 
-run() takes the number of simulated as first parameter, and the differential, or step, used for the Euler approximation of the ODE.
+run() takes the number of simulated as first parameter, and the differential, or step, used for the Euler approximation of the ODE. 
 
 If a plot is wanted, simply run the following:
 
@@ -23,4 +23,4 @@ df, fig = sim.run(31, .1)
 
 ```
 
-This will display a plot of all variables, which can be further customized using the default boolean parameters.
+This will display a plot of all variables, which can be further customized using the default boolean parameters that represent all of the variables within the model.
