### 2021-05-24 f8460b1
Added some more to paper.md
