### 2021-05-24 #3247 @mjacob1002
Hi all,

Sorry for the late replies - I'm currently in the middle of exam season and have been preoccupied the past couple of days.

@joeweaver, I think I satisfied the two issues that brought up. Please let me know if my changes fix the problem.

@DARSakthi, thanks for the feedback! I think I fixed the documentation issues. For the SEIR and SIRS documentation issues, however, could you please give the name of the file that you found the issues with? I personally couldn't find the issues with the 'beta' in the deterministic SIRS documentation file. I made sure to say that run() takes parameters in the deterministic docs. Also, for the SIS, I made sure to mention that it was a class, although I hope that the example included in the documentation helps to clear that up as well.

For the paper, I made sure to switch the Summary and Statement of Need, as well as bolster the new Statement of Need.

Please let me know if there is anything else I should do to improve my repo!
