### 2021-05-24 #3247 @mjacob1002
Also @will-rowe, should I mention that this project received no funding, or should I remove that from the paper?
