### 2021-05-24 #3247 @DARSakthi
That's great -- I'll be sure to take a look again and give (hopefully final) remarks. New documentation looks much better

it seems as though [https://github.com/mjacob1002/Eir/blob/master/docs/Deterministic/SIRS.md](https://github.com/mjacob1002/Eir/blob/master/docs/Deterministic/SIRS.md) and SEIR have beta in it -- I recall not seeing this and being confused when the model threw an error, but I'm afraid I don't recall what file I was looking at. If this is intended as the documentation users will look at then this is a non-issue

Otherwise, this looks great. Best of luck on your exams

Cheers

DS
