### 2021-05-30 mjacob1002/Eir#10 @mjacob1002
Hi @joeweaver, just following up to make sure that the commit to the paper was more in line with what you wanted.
