### 2021-05-30 mjacob1002/Eir#10 @joeweaver
Hi,

The earlier email flew under my radar somehow I'll be checking out the
changes on Monday.

--Joe

On Sun, May 30, 2021 at 5:54 AM Mathew Jacob ***@***.***>
wrote:

> Hi @joeweaver <https://github.com/joeweaver>, just following up to make
> sure that the commit to the paper was more in line with what you wanted.
>
