### 2021-06-01 #3247 @DARSakthi
Closing comments:

In the end of the summary, I would still specify what 'certain assumptions' means. It currently isn't a very meaningful sentence. Perhaps something like 'used by epidemiologists and researchers who are interested in understanding the dynamics of epidemics when operating under different modelling assumptions, by evaluating different models in ```Eir```' (I assume this is what you mean)

So far as I can discern, the statement of need is no longer a literature review instead of a statement of need. This is a substantial improvement. I have the same feedback for the related packages section.

I would encourage the edit I suggested in the first point. Then, I would find the paper satisfactory, and recommend publication immediately after. Note all boxes are checked and this is a mere cosmetic revision.

By the way, I think the new paper was never compiled into a PDF. I can only access the .md file.

Cheers

DS
