### 2021-06-02 mjacob1002/Eir#10 @joeweaver
The updated docs clear up what the tests are doing.
