### 2021-06-02 mjacob1002/Eir#9 @joeweaver
The updated docs clear up what the tests are doing.
