### 2021-06-02 mjacob1002/Eir#10 @joeweaver
Removed a comment meant for another issue.  The updated paper places the software in the context of the other packages.
