### 2021-06-03 #3247 @will-rowe
Thanks so much @DARSakthi and @joeweaver - these are excellent reviews. My apologies for a slow response but I've been away recently.

@mjacob1002 - it all looks good to me and I'm happy to move toward acceptance. I've got no issue with the "no funding" statement and I can't find anything about it in the guidelines. That being said, an EiC may have better guidance but we'll cross that bridge if we come to it.
