### 2021-06-03 #3247 @will-rowe
Okay - nearly there. Please address the following:

ln 10: dont' -> don't
ln25: should probably be "... the work of Mühlpfordt et al. (Mühlpfordt, 2020)"
ln31: is there a reference or DOI for this study?
ln44: researcher -> researchers
ln45: add space after Markov chains
