### 2021-06-06 b33dc33
Fixed title of plot of periodic ICUV
