### 2021-06-06 #3247 @mjacob1002
Hi @will-rowe,

Sorry for the delay. I made all of the changes you suggested except for the third issue because the study is still ongoing with no preprint currently, so there is nothing to cite at the moment. Maybe I can add it after the preprint is released.

The version number of the release is v1.0.2. The archive can be found on Zenodo: https://zenodo.org/record/4905077#.YL1EcJNKhO4.
