### 2021-06-08 #3247 @kthyng
Version looks good.

@mjacob1002 can you update the title and author list in the Zenodo archive metadata to exactly match your JOSS paper?
