### 2021-06-08 #3247 @mjacob1002
Hi @kthyng,

I changed the affiliation in the Zenodo to be RxCovea, New York University. I believe it should know match the JOSS paper, but let me know if there are other specific things that need to be changed.
