### 2021-06-08 #3247 @kthyng
Paper comments:
* Your affiliation is pretty empty. Can you make this more standardized?
* 2nd paragraph: there are a bunch of acronyms without being written out. Is there a reason you don't explain what they are?
* "The ICU models..." sentence ends with the period in the wrong place and incorrect spacing.
* In Statement of need: spacing with the Fujie reference is off.
* Epidemics on Networks: the reference should not be parenthetical, so remove the [] around the reference.
* Jenness reference doesn't capitalize R.
* Generally, please read through your paper draft and polish it. The next step is the final paper publication.
