### 2021-06-08 #3247 @kthyng
> I changed the affiliation in the Zenodo to be RxCovea, New York University. I believe it should know match the JOSS paper, but let me know if there are other specific things that need to be changed.

Please change the title and author list in the Zenodo archive metadata to exactly match your JOSS paper.
