### 2021-06-08 99cd92c
Fixed minor typos and the title problem in the .bib file.
