+++ paper/paper.bib
@@ -3,7 +3,7 @@ author = {Kermack, W.O. and McKendrick, A.G.A.},
 year = {1927},
 month = {01},
 pages = {700-721},
-title = {A Contribution to the Mathematical Theory of Epidemics},
+title = {{A Contribution to the Mathematical Theory of Epidemics}},
 volume = {115},
 journal = {Proc. Roy. Soc. Edinburgh},
 doi = {10.1098/rspa.1927.0118}
@@ -14,7 +14,7 @@ author = {Fujie, Ryo and Odagaki, Takashi},
 year = {2007},
 month = {02},
 pages = {843-852},
-title = {Effects of superspreaders in spread of epidemic},
+title = {{Effects of superspreaders in spread of epidemic}},
 volume = {374},
 journal = {Physica A-statistical Mechanics and Its Applications - PHYSICA A},
 doi = {10.1016/j.physa.2006.08.050}
@@ -25,7 +25,7 @@ author = {Allen, Linda},
 year = {2017},
 month = {03},
 pages = {},
-title = {A primer on stochastic epidemic models: Formulation, numerical simulation, and analysis},
+title = {{A primer on stochastic epidemic models: Formulation, numerical simulation, and analysis}},
 volume = {2},
 journal = {Infectious Disease Modelling},
 doi = {10.1016/j.idm.2017.03.001}
@@ -33,7 +33,7 @@ doi = {10.1016/j.idm.2017.03.001}
 
 @misc{ICU,
   author = {Tillmann Mühlpfordt},
-  title = {SEIR Model with Stochastic Uncertainties -- A Polynomial Chaos Approach},
+  title = {{SEIR Model with Stochastic Uncertainties -- A Polynomial Chaos Approach}},
   year = {2020},
   publisher = {GitHub},
   journal = {GitHub repository},
@@ -49,13 +49,13 @@ doi = {10.1016/j.idm.2017.03.001}
   number = {44},
   pages = {1731},
   author = {Joel C. Miller and Tony Ting},
-  title = {EoN (Epidemics on Networks): a fast, flexible Python package for simulation, analytic approximation, and analysis of epidemics on networks},
+  title = {{EoN (Epidemics on Networks): a fast, flexible Python package for simulation, analytic approximation, and analysis of epidemics on networks}},
   journal = {Journal of Open Source Software}
 }
 
 @article{EpiModel,
    author = {Samuel M. Jenness and Steven M. Goodreau and Martina Morris},
-   title = {EpiModel: An R Package for Mathematical Modeling of Infectious Disease over Networks},
+   title = {{EpiModel: An R Package for Mathematical Modeling of Infectious Disease over Networks}},
    journal = {Journal of Statistical Software, Articles},
    volume = {84},
    number = {8},
