### 2021-06-10 b550723
Added the full meaning of compartmental model acronyms.
