### 2021-06-10 47c7559
Paper details 2.0
