+++ paper/paper.md
@@ -24,9 +24,9 @@ bibliography: paper.bib
 
 # Summary
 
-` Eir ` is a Python package that simulates epidemics with various assumptions and compartmental models. By mainly using spatial models, ` Eir ` allows for better modeling of the spread of an epidemic in smaller communities and localities. There are four different spatial models used: the Strong Infectious Model [@Fujie&Odagaki], Hub Model [@Fujie&Odagaki], the Random Movement Model, and the Periodic Mobility Model. While the Strong Infectious Model and Hub Model are static models, where no movement pattern is simulated, the Random Movement and Periodic Mobility Model simulates the movement pattern of people using different assumptions. The Random Movement Model assumes that an individual is equally likely to move in any direction for a random distance, picked from a user-defined normal distribution. The Periodic Mobility Model assumes that individuals move in a circle to capture the routine nature of people's movements. Therefore, after randomly generating locations, the model calculates a center of a circle and individuals move along that circle. With all of these spatial models, researchers can use the `Simul_Details` objects to interact with the simulation data and get a deeper understanding of the simulation's dynamics, such as transmission chains, state histories, etc. The data from these simulations can be plotted using matplotlib or be returned in the form of a pandas DataFrame. It is important to note that these models are stochastic and are discrete-time Markov chains. 
+` Eir ` is a Python package that simulates epidemics with various assumptions and compartmental models. By mainly using spatial models, ` Eir ` allows for better modeling of the spread of an epidemic in smaller communities and localities. There are four different spatial models used: the Strong Infectious Model [@Fujie&Odagaki], Hub Model [@Fujie&Odagaki], the Random Movement Model, and the Periodic Mobility Model. While the Strong Infectious Model and Hub Model are static models, where no movement pattern is simulated, the Random Movement and Periodic Mobility Model simulates the movement pattern of people using different assumptions. The Random Movement Model assumes that an individual is equally likely to move in any direction for a random distance, generated from a user-defined normal distribution. The Periodic Mobility Model assumes that individuals move in a circle to capture the routine nature of people's movements. Therefore, after randomly generating locations, the model calculates a center of a circle and individuals move along that circle. With all of these spatial models, researchers can use the `Simul_Details` objects to interact with the simulation data and get a deeper understanding of the simulation's dynamics, such as transmission chains, state histories, etc. The data from these simulations can be plotted using Matplotlib or be returned in the form of a pandas DataFrame. It is important to note that these models are stochastic and are discrete-time Markov chains. 
 
-` Eir ` leverages many different compartmental models in addtition to different spatial models in order to give researchers flexibility and robustness when simulating the spread of an epidemic. The models include but are not limited to: Susceptible-Infected-Recovered(SIR), Susceptible-Infected-Recovered-Dead(SIRD), Susceptible-Exposed-Infected-Recovered(SEIR), Susceptible-Exposed-Infected-Recovered-Vaccinated(SEIRV), Susceptible-Infected-Recovered-Susceptible-Vaccinated(SIRSV), and ICU models. The ICU models, in particular, are based on the work of Mühlpfordt et al[@ICU]. Additionally, for any model that includes vaccinations, there is a timeDelay variable that allows the user to delay vaccine distribution by a particular amount of days. 
[…]
