### 2021-06-10 #3247 @mjacob1002
Hi @kthyng ,

Thanks for your feedback. I've gone through the paper and cleaned up phrasing and syntax errors, but please let me know if I missed any. Additionally, I changed the Zenodo and author list to match the JOSS paper.

Let me know if there is anything else I should modify in the Zenodo archive or the paper.

- Mathew
