### 2021-06-10 #3247 @mjacob1002
Thanks so much @kthyng, @will-rowe, @joeweaver, and @DARSakthi for your hardwork in reviewing my submission - it means a lot (this is my first time publishing, and it has been a wonderful experience!)

Quick follow up question: on the JOSS website, it says that there is a 404 error when trying to display the paper - will this clear up or is something wrong on my end?
